add test for add user to db function
