+++ tests/test_authentication.py
@@ -3,6 +3,10 @@
 from pytest_mock import MockerFixture
 
 from insilicoserver.user_authentication import authenticate_username
+from insilicoserver.mongo_interface import USERS_COLLECTION, add_users_to_db
+
+from .test_mongo_interface import get_database
+from .utils_test import PATH_TEST
 
 
 class MockReply(dict):
@@ -26,3 +30,14 @@ def test_correct_token(mocker: MockerFixture):
     username = authenticate_username("validtoken")
 
     assert username == "felipeZ"
+
+
+def test_add_user_to_db():
+    """Check that some users are properly added in the database."""
+    path_users = PATH_TEST / "users.txt"
+
+    try:
+        db = get_database()
+        add_users_to_db(db, path_users)
+    finally:
+        db.drop_collection(USERS_COLLECTION)
